+++ CWE/code_system/codes/include/race/code_karatemedal.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../../../cwe_code.h"
+
+class KarateGiveMedal : public Code {
+public:
+	using Code::Code;
+
+	void Init() override;
+};
